Fix redundant error handling
