+++ src/errors/handlers.ts
@@ -1,5 +1,24 @@
 import { Request, Response, NextFunction } from "express";
-import { MethodNotAllowedError } from "./http";
+import { NotFoundError, MethodNotAllowedError } from "./http";
+
+/**
+ * Class to handle 404 Not Found errors
+ */
+export class NotFoundHandler {
+  /**
+   * Middleware to handle 404 Not Found errors
+   * @param req - Express request object
+   * @param res - Express response object
+   * @param next - Express next function
+   */
+  static handle(req: Request, _res: Response, next: NextFunction) {
+    const error = new NotFoundError(`Not found: ${req.originalUrl}`, {
+      url: req.originalUrl,
+      method: req.method,
+    });
+    next(error);
+  }
+}
 
 /**
  * Class to handle 405 Method Not Allowed errors
@@ -22,6 +41,3 @@ export class MethodNotAllowedHandler {
     next(error);
   }
 }
-
-// For backward compatibility
-export default MethodNotAllowedHandler.handle;
