+++ src/errors/index.ts
@@ -7,5 +7,10 @@ export * from "./types";
 // Error handlers
 export { ErrorHandler } from "./errorHandler";
 export { AsyncErrorBoundary } from "./asyncErrorBoundary";
-export { MethodNotAllowedHandler } from "./methodNotAllowed";
-export { NotFoundHandler } from "./notFound";
+export { NotFoundHandler, MethodNotAllowedHandler } from "./handlers";
+
+// For backward compatibility
+export { default as errorHandler } from "./errorHandler";
+export { default as asyncErrorBoundary } from "./asyncErrorBoundary";
+export { NotFoundHandler as notFound } from "./handlers";
+export { MethodNotAllowedHandler as methodNotAllowed } from "./handlers";
